+++ poem-openapi/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [1.0.46] 2021-12-10
+
+- Change the default renaming rule of enum items from `ScreamingSnake` to `Pascal`.
+
 # [1.0.45] 2021-12-09
 
 - Remove the `desc` attribute of the response header in `ApiResponse` macro, and use rustdoc to add the header description.